+++ tests/func/parsing/error_reporting/test_parsing_errors.py
@@ -1,7 +1,7 @@
 import pytest  # noqa
 import os
 from pathlib import Path
-from parglare import Grammar, Parser, GLRParser, ParseError
+from parglare import Grammar, Parser, GLRParser, SyntaxError
 from ...grammar.expression_grammar import get_grammar
 from ...utils import output_cmp
 
@@ -15,7 +15,7 @@ def test_grammar_in_error(parser_class):
     grammar = get_grammar()
     p = parser_class(grammar)
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("id+id*+id")
 
     assert e.value.grammar is grammar
@@ -26,7 +26,7 @@ def test_glr_last_heads_in_error():
     grammar = get_grammar()
     p = GLRParser(grammar)
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("id+id*+id")
 
     assert len(e.value.last_heads) == 1
@@ -38,7 +38,7 @@ def test_invalid_input(parser_class):
     grammar = get_grammar()
     p = parser_class(grammar)
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("id+id*+id")
 
     assert e.value.location.start_position == 6
@@ -58,7 +58,7 @@ def test_premature_end(parser_class):
     grammar = get_grammar()
     p = parser_class(grammar)
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("id+id*")
 
     assert e.value.location.start_position == 6
@@ -84,7 +84,7 @@ def test_ambiguous_glr():
     g = Grammar.from_string(grammar)
     parser = GLRParser(g)
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         parser.parse("1 + 2 * 3 / 5")
 
     assert e.value.location.start_position == 10
@@ -98,7 +98,7 @@ def test_line_column(parser_class):
     grammar = get_grammar()
     p = parser_class(grammar)
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("""id + id * id + id + error * id""")
 
     loc = e.value.location
@@ -106,7 +106,7 @@ def test_line_column(parser_class):
     assert loc.line == 1
     assert loc.column == 20
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("""id + id * id + id + error * id
 
         """)
@@ -115,7 +115,7 @@ def test_line_column(parser_class):
     assert loc.line == 1
     assert loc.column == 20
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("""
 
 id + id * id + id + error * id""")
@@ -124,7 +124,7 @@ id + id * id + id + error * id""")
     assert loc.line == 3
     assert loc.column == 20
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse("""
 
 id + id * id + id + error * id
@@ -145,7 +145,7 @@ def test_parser_output(parser_class):
     input_file = Path(os.path.dirname(__file__), 'parsing_errors.input')
     err_file = Path(os.path.dirname(__file__), 'parsing_errors.err')
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse_file(input_file)
 
     output_cmp(err_file, str(e.value))
@@ -159,7 +159,7 @@ def test_file_name(parser_class):
 
     input_file = Path(os.path.dirname(__file__), 'parsing_errors.input')
 
-    with pytest.raises(ParseError) as e:
+    with pytest.raises(SyntaxError) as e:
         p.parse_file(str(input_file))
 
     assert 'parsing_errors.input' in str(e.value)