+++ CHANGELOG.md
@@ -22,12 +22,13 @@ backward incompatible changes will start to apply when the projects goes 1.0
 - Support for Python 3.13 and 3.14.
 
 ## Fixed
-- respect `--debug` flag in `pglr parse`
-- ignore `KEYWORD` terminal getting all lookahead tokens on parse failure
+- respect `--debug` flag in `pglr parse`.
+- ignore `KEYWORD` terminal getting all lookahead tokens on parse failure.
 
 ### Changed
 - Moved to [uv](https://docs.astral.sh/uv/) for dependency/project management.
 - Improved error reporting.
+- Class name `ParseError` changed to `SyntaxError`.
 
 
 ## [0.18.0] (released: 2024-02-23)